Version 1.1 - Affichage en cas d'éventuel remboursement pour un cours à venir - Ajout d'une feneêtre de payement - Modification problème d'affichage pour les cours - Ajout d'éléments dans le tableau cours - Les cours collectifs n'ont plus de limitation de temps - Ajout des jours pour les cours particuliers - Suppression de classes inutilisées dans le programme - Affichage uniquement des réservations à venir - Mise à jour du moniteur selon la semaine - Tentative d'ajout d'une progress bar - annulée

+++ Programme/Ski_Java/src/be/mousty/accessToDao/CoursATD.java
@@ -42,7 +42,7 @@ public class CoursATD {
 	public boolean 				update				(Cours c) 		{ return CoursDAO.update(c); 	}
 	public Cours 				find				(int id) 		{ return CoursDAO.find(id); 	} 
 	public ArrayList<Cours> 	getListC			() 				{ return CoursDAO.getList(); 	} 
-	public String 				calculerPlaceCours 	(int idCours, int idSemaine, int idMoniteur) 
+	public String 				calculerPlaceCours 	(int idCours, long idSemaine, int idMoniteur) 
 	{ return CoursDAO.calculerPlaceCours(idCours, idSemaine, idMoniteur); }
 	public ArrayList<Cours> 	getListCoursSelonId	(int idMoniteur){ return CoursDAO.getMyListSelonID(idMoniteur, -1, -1, ""); 	}
